+++ tests/test_components.py
@@ -36,6 +36,7 @@ def test_component(component):
         'save_tex': True,
         'show_tex_output': SHOW_TEX_OUTPUT,
         'no_pdf': False,
+        'use_packages': None, # default to texlite/config/packages.json
     }
 
     # execute main